Move GroupHeading out to separate file in src/Query/
